requirements.txt
atlassian-python-api>=3.41.0
certifi>=2023.7.0
charset-normalizer>=3.3.0
Deprecated>=1.2.0
idna>=3.4.0
oauthlib>=3.2.0
requests>=2.31.0
requests-oauthlib>=1.3.0
six>=1.16.0
urllib3<2.0.0
wrapt>=1.15.0
python-dotenv>=1.0.0
click>=8.0.0
pydantic>=2.0.0
